Palette KD-tree: Fix compilation speed
Tricks the compiler into skipping expensive
`uninit var analysis` (`-Wmaybe-uninitialized`) by using
a struct with state rather than separate variables for `best` /
`bestDiff`.

This has no performance impact.

Also optimizes lookup a bit further and moves some code that
does not need to be inlined to the cpp file.

```
Benchmark                                              Time             CPU      Time Old      Time New       CPU Old       CPU New
-----------------------------------------------------------------------------------------------------------------------------------
BM_GenerateBlendedLookupTable_pvalue                 0.0002          0.0002      U Test, Repetitions: 10 vs 10
BM_GenerateBlendedLookupTable_mean                  +0.0237         +0.0237       2092090       2141601       2091732       2141291
BM_GenerateBlendedLookupTable_median                +0.0237         +0.0237       2092104       2141662       2091669       2141319
BM_GenerateBlendedLookupTable_stddev                -0.6414         -0.5834           664           238           538           224
BM_GenerateBlendedLookupTable_cv                    -0.6497         -0.5930             0             0             0             0
BM_BuildTree_pvalue                                  0.0002          0.0002      U Test, Repetitions: 10 vs 10
BM_BuildTree_mean                                   +0.0410         +0.0410          4495          4679          4494          4678
BM_BuildTree_median                                 +0.0403         +0.0402          4494          4675          4493          4674
BM_BuildTree_stddev                                 +0.9515         +0.9359             7            14             7            14
BM_BuildTree_cv                                     +0.8746         +0.8596             0             0             0             0
BM_FindNearestNeighbor_pvalue                        0.0002          0.0002      U Test, Repetitions: 10 vs 10
BM_FindNearestNeighbor_mean                         -0.0399         -0.0398    1964257108    1885966812    1963954917    1885694336
BM_FindNearestNeighbor_median                       -0.0397         -0.0396    1963969748    1886074435    1963650984    1885803182
BM_FindNearestNeighbor_stddev                       -0.3380         -0.3443       1217360        805946       1225442        803469
BM_FindNearestNeighbor_cv                           -0.3105         -0.3171             0             0             0             0
OVERALL_GEOMEAN                                     +0.0077         +0.0077             0             0             0             0
```
